requirements.txt
streamlit
pandas
plotly
pyyaml
python-dotenv
streamlit-aggrid